+++ cmd/uniflow/main.go
@@ -13,9 +13,9 @@ import (
 	"github.com/siyul-park/uniflow/pkg/hook"
 	"github.com/siyul-park/uniflow/pkg/scheme"
 	"github.com/siyul-park/uniflow/pkg/storage"
-	"github.com/siyul-park/uniflow/third_party/controllx"
-	"github.com/siyul-park/uniflow/third_party/networkx"
-	"github.com/siyul-park/uniflow/third_party/systemx"
+	"github.com/siyul-park/uniflow/plugin/pkg/controllx"
+	"github.com/siyul-park/uniflow/plugin/pkg/networkx"
+	"github.com/siyul-park/uniflow/plugin/pkg/systemx"
 	"github.com/spf13/viper"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
