+++ plugin/pkg/controllx/switch_test.go
